made header & footer a child component of mainpage, restructured component folder layout
